+++ infra/postgres/migrations/0042_skill_quality_warning.down.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+DROP TABLE IF EXISTS skill_quality_warning;
+DROP TABLE IF EXISTS skill_quality_observation;
+
+COMMIT;